+++ qacode/core/webs/__init__.py
@@ -3,16 +3,7 @@
 
 
 from qacode.core.webs import controls
-from qacode.core.webs import css_properties
-from qacode.core.webs import html_attrs
-from qacode.core.webs import html_tags
 from qacode.core.webs import pages
 
 
-__all__ = [
-    'controls',
-    'css_properties',
-    'html_attrs',
-    'html_tags',
-    'pages',
-]
+__all__ = ['controls', 'pages']